requirements.txt
absl-py==0.12.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.0
astunparse==1.6.3
autopep8==1.6.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
branca==0.4.2
Brlapi==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2
django-bootstrap4==2.3.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
ffprobe-python==1.0.3
filelock==3.0.12
flatbuffers==1.12
folium==0.12.1
fpdf2==2.4.5
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.14.0
idna==2.10
importlib-metadata==2.1.1
Jinja2==3.0.1
joblib==1.0.1
jsonfield==3.1.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
kmeans1d==0.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
meson==0.53.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas==1.3.4
paramiko==2.6.0
pdfminer.six==20200517
pep8==1.7.1
pexpect==4.6.0
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycryptodome==3.10.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymssql==2.2.1
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
PyPDF4==1.27.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.5.34
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
systemd-python==234
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
Wand==0.6.6
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
xkit==0.0.0
youtube-dl==2021.6.6
zipp==3.5.0
